refactor: extract common code between v1 and v1beta1
