+++ crd-generator/src/main/java/io/fabric8/crd/generator/v1/JsonSchema.java
@@ -15,66 +15,15 @@
  */
 package io.fabric8.crd.generator.v1;
 
-import io.fabric8.crd.generator.utils.Types;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaProps;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaPropsBuilder;
-import io.fabric8.kubernetes.client.CustomResource;
-import io.sundr.builder.internal.functions.TypeAs;
-import io.sundr.codegen.functions.ClassTo;
-import io.sundr.codegen.model.ClassRef;
-import io.sundr.codegen.model.PrimitiveRefBuilder;
 import io.sundr.codegen.model.Property;
 import io.sundr.codegen.model.TypeDef;
-import io.sundr.codegen.model.TypeRef;
-import io.sundr.codegen.utils.TypeUtils;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 
-public class JsonSchema {
+public class JsonSchema extends io.fabric8.crd.generator.JsonSchema<JSONSchemaProps, JSONSchemaPropsBuilder> {
 
-  private static final TypeDef CUSTOM_RESOURCE = ClassTo.TYPEDEF.apply(CustomResource.class);
-  private static final TypeDef BOOLEAN = ClassTo.TYPEDEF.apply(Boolean.class);
-  private static final TypeDef STRING = ClassTo.TYPEDEF.apply(String.class);
-  private static final TypeDef INT = ClassTo.TYPEDEF.apply(Integer.class);
-  private static final TypeDef LONG = ClassTo.TYPEDEF.apply(Long.class);
-  private static final TypeDef DOUBLE = ClassTo.TYPEDEF.apply(Double.class);
-  private static final TypeDef DATE = ClassTo.TYPEDEF.apply(Date.class);
-  private static final TypeDef OPTIONAL = ClassTo.TYPEDEF.apply(Optional.class);
-
-  private static final TypeRef BOOLEAN_REF = BOOLEAN.toReference();
-  private static final TypeRef STRING_REF = STRING.toReference();
-  private static final TypeRef INT_REF = INT.toReference();
-  private static final TypeRef LONG_REF = LONG.toReference();
-  private static final TypeRef DOUBLE_REF = DOUBLE.toReference();
-  private static final TypeRef DATE_REF = DATE.toReference();
-
-  private static final TypeRef P_INT_REF = new PrimitiveRefBuilder().withName("int").build();
-  private static final TypeRef P_LONG_REF = new PrimitiveRefBuilder().withName("long").build();
-  private static final TypeRef P_DOUBLE_REF = new PrimitiveRefBuilder().withName("double").build();
-  private static final TypeRef P_BOOLEAN_REF = new PrimitiveRefBuilder().withName("boolean").build();
-  private static final TypeRef OPTIONAL_REF = OPTIONAL.toReference();
-
-  public static final Map<TypeRef, String> TYPE_MAP = new HashMap<>();
-  static {
-    TYPE_MAP.put(STRING_REF, "string");
-    TYPE_MAP.put(DATE_REF, "string");
-    TYPE_MAP.put(INT_REF, "integer");
-    TYPE_MAP.put(P_INT_REF, "integer");
-    TYPE_MAP.put(LONG_REF, "number");
-    TYPE_MAP.put(P_LONG_REF, "number");
-    TYPE_MAP.put(DOUBLE_REF, "number");
-    TYPE_MAP.put(P_DOUBLE_REF, "number");
-    TYPE_MAP.put(BOOLEAN_REF, "boolean");
-    TYPE_MAP.put(P_BOOLEAN_REF, "boolean");
-  }
+  private static final JsonSchema instance = new JsonSchema();
 
   /**
    * Creates the JSON schema for the particular {@link TypeDef}.
@@ -84,59 +33,42 @@ public class JsonSchema {
    * @return The schema.
    */
   public static JSONSchemaProps from(TypeDef definition, String ... ignore) {
-    JSONSchemaPropsBuilder builder = new JSONSchemaPropsBuilder();
-    builder.withType("object");
-    Set<String> ignores  = ignore.length > 0 ? new LinkedHashSet<>(Arrays.asList(ignore)) : Collections.emptySet();
-    List<String> required = new ArrayList<>();
-
-
-    for (Property property : Types.allProperties(definition)) {
-      if (property.isStatic()) {
-        continue;
-      }
-
-      if (ignores.contains(property.getName())) {
-        continue;
-      }
-
-      JSONSchemaProps schema = from(property.getTypeRef());
-      if (property.getAnnotations()
-          .stream()
-          .anyMatch(a -> a.getClassRef().getFullyQualifiedName().equals("javax.validation.constraints.NotNull"))) {
-        required.add(property.getName());
-      }
-
-      if (schema != null) {
-        builder.addToProperties(property.getName(), from(property.getTypeRef()));
-      }
-    }
-    builder.withRequired(required);
-    return builder.build();
+    return instance.internalFrom(definition, ignore);
   }
 
-  public static JSONSchemaProps from(TypeRef typeRef) {
-    //1. Handle Collections and Arrays
-    if (typeRef.getDimensions() > 0 || TypeUtils.isCollection(typeRef)) {
-      return new JSONSchemaPropsBuilder()
-          .withType("array")
-          .withNewItems()
-          .withSchema(from(TypeAs.combine(TypeAs.UNWRAP_ARRAY_OF, TypeAs.UNWRAP_COLLECTION_OF).apply(typeRef)))
-          .and()
-          .build();
-      //2. Handle Standard Types
-    } else if (TYPE_MAP.containsKey(typeRef)) {
-      return new JSONSchemaPropsBuilder()
-          .withType(TYPE_MAP.get(typeRef))
-          .build();
-      //3. Handle Optionals
-    } else if (TypeUtils.isOptional(typeRef)) {
-      return from(TypeAs.UNWRAP_OPTIONAL_OF.apply(typeRef));
-      //4. Handle complex types
-    } else if (typeRef instanceof ClassRef) {
-      ClassRef classRef = (ClassRef) typeRef;
-      TypeDef def = classRef.getDefinition();
-      return from(def);
+  @Override
+  public JSONSchemaPropsBuilder newBuilder() {
+    final JSONSchemaPropsBuilder builder = new JSONSchemaPropsBuilder();
+    builder.withType("object");
+    return builder;
+  }
+
+  @Override
+  public void addProperty(Property property, JSONSchemaPropsBuilder builder, JSONSchemaProps schema) {
+    if (schema != null) {
+      builder.addToProperties(property.getName(), schema);
     }
-    return null;
+  }
+
+  @Override
+  public JSONSchemaProps build(JSONSchemaPropsBuilder builder, List<String> required) {
+    return builder.withRequired(required).build();
+  }
+
+  @Override
+  protected JSONSchemaProps collectionProperty(JSONSchemaProps schema) {
+    return new JSONSchemaPropsBuilder()
+      .withType("array")
+      .withNewItems()
+      .withSchema(schema)
+      .and()
+      .build();
+  }
+
+  @Override
+  protected JSONSchemaProps singleProperty(String typeName) {
+    return new JSONSchemaPropsBuilder()
+      .withType(typeName)
+      .build();
   }
 }
